End Libft Without Bonus

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: hbourgeo <hbourgeo@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/14 20:50:09 by hbourgeo          #+#    #+#             */
-/*   Updated: 2021/10/14 21:43:26 by hbourgeo         ###   ########.fr       */
+/*   Updated: 2021/10/24 20:19:45 by hbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*sub_str;
-	int		n;
+	size_t	n;
 
 	n = 0;
 	sub_str = ft_calloc(len + 1, sizeof(char));
